Revert "Created /users api endpoints and removed /auth endpoints"
This reverts commit 5962ed1c946c45f935ad2e57e2992816883b1006.
