+++ api/app.py
@@ -1,5 +1,5 @@
 from flask import Flask
-from api.views.users import users_bp
+from api.views.auth import auth_bp
 from api.models import db
 
 from api import config
@@ -9,7 +9,7 @@ app.config.from_object(config)
 
 db.init_app(app)
 
-app.register_blueprint(users_bp, url_prefix='/users')
+app.register_blueprint(auth_bp, url_prefix='/auth')
 
 if __name__ == '__main__':
     app.run(debug=True)
